tests/ update tests for fetcher
